refactor: 나와 관련된 운동 컨트롤러 분리
